requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2 
FPDF 
streamlit-quill 
fpdf2